Network.AWS.DirectoryService.AddIPRoutes

Description

If the DNS server for your on-premises domain uses a publicly addressable IP address, you must add a CIDR address block to correctly route traffic to and from your Microsoft AD on Amazon Web Services. AddIpRoutes adds this address block. You can also use AddIpRoutes to facilitate routing traffic that uses public IP ranges from your Microsoft AD on AWS to a peer VPC.

Before you call AddIpRoutes , ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the AddIpRoutes operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference .

Creating a Request

addIPRoutes

Arguments

:: Text
airDirectoryId

-> AddIPRoutes

Creates a value of AddIPRoutes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

airUpdateSecurityGroupForDirectoryControllers - If set to true, updates the inbound and outbound rules of the security group that has the description: "AWS created security group for directory ID directory controllers." Following are the new rules: Inbound: * Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.00 * Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.00 * Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.00 * Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.00 * Type: All ICMP, Protocol: All, Range: NA, Source: 0.0.0.00 Outbound: * Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0 These security rules impact an internal network interface that is not exposed publicly.

airDirectoryId - Identifier (ID) of the directory to which to add the address block.

airIPRoutes - IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your on-premises domain.

data AddIPRoutes

See: addIPRoutes smart constructor.

Request Lenses

airUpdateSecurityGroupForDirectoryControllers :: Lens' AddIPRoutes (Maybe Bool)

If set to true, updates the inbound and outbound rules of the security group that has the description: "AWS created security group for directory ID directory controllers." Following are the new rules: Inbound: * Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.00 * Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.00 * Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.00 * Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.00 * Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.00 * Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.00 * Type: All ICMP, Protocol: All, Range: NA, Source: 0.0.0.00 Outbound: * Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0 These security rules impact an internal network interface that is not exposed publicly.

airDirectoryId :: Lens' AddIPRoutes Text

Identifier (ID) of the directory to which to add the address block.

airIPRoutes :: Lens' AddIPRoutes [IPRoute]

IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your on-premises domain.

Destructuring the Response

addIPRoutesResponse

Arguments

:: Int
airrsResponseStatus

-> AddIPRoutesResponse

Creates a value of AddIPRoutesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

airrsResponseStatus - -- | The response status code.

data AddIPRoutesResponse

See: addIPRoutesResponse smart constructor.

Response Lenses

airrsResponseStatus :: Lens' AddIPRoutesResponse Int

- | The response status code.