Network.AWS.DirectoryService.CreateAlias

Description

Creates an alias for a directory and assigns the alias to the directory. The alias is used to construct the access URL for the directory, such as http://<alias>.awsapps.com .

Important: After an alias has been created, it cannot be deleted or reused, so this operation should only be used when absolutely necessary.

Creating a Request

createAlias

Arguments

:: Text
caDirectoryId

-> Text
caAlias

-> CreateAlias

Creates a value of CreateAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caDirectoryId - The identifier of the directory for which to create the alias.

caAlias - The requested alias. The alias must be unique amongst all aliases in AWS. This operation throws an EntityAlreadyExistsException error if the alias already exists.

data CreateAlias

Contains the inputs for the CreateAlias operation.

See: createAlias smart constructor.

Request Lenses

caDirectoryId :: Lens' CreateAlias Text

The identifier of the directory for which to create the alias.

caAlias :: Lens' CreateAlias Text

The requested alias. The alias must be unique amongst all aliases in AWS. This operation throws an EntityAlreadyExistsException error if the alias already exists.

Destructuring the Response

createAliasResponse

Arguments

:: Int
carsResponseStatus

-> CreateAliasResponse

Creates a value of CreateAliasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsDirectoryId - The identifier of the directory.

carsAlias - The alias for the directory.

carsResponseStatus - -- | The response status code.

data CreateAliasResponse

Contains the results of the CreateAlias operation.

See: createAliasResponse smart constructor.

Response Lenses

carsDirectoryId :: Lens' CreateAliasResponse (Maybe Text)

The identifier of the directory.

carsAlias :: Lens' CreateAliasResponse (Maybe Text)

The alias for the directory.

carsResponseStatus :: Lens' CreateAliasResponse Int

- | The response status code.