Network.AWS.DirectoryService.CreateComputer

Description

Creates a computer account in the specified directory, and joins the computer to the directory.

Creating a Request

createComputer

Arguments

:: Text
ccDirectoryId

-> Text
ccComputerName

-> Text
ccPassword

-> CreateComputer

Creates a value of CreateComputer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccComputerAttributes - An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

ccOrganizationalUnitDistinguishedName - The fully-qualified distinguished name of the organizational unit to place the computer account in.

ccDirectoryId - The identifier of the directory in which to create the computer account.

ccComputerName - The name of the computer account.

ccPassword - A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

data CreateComputer

Contains the inputs for the CreateComputer operation.

See: createComputer smart constructor.

Request Lenses

ccComputerAttributes :: Lens' CreateComputer [Attribute]

An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

ccOrganizationalUnitDistinguishedName :: Lens' CreateComputer (Maybe Text)

The fully-qualified distinguished name of the organizational unit to place the computer account in.

ccDirectoryId :: Lens' CreateComputer Text

The identifier of the directory in which to create the computer account.

ccComputerName :: Lens' CreateComputer Text

The name of the computer account.

ccPassword :: Lens' CreateComputer Text

A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

Destructuring the Response

createComputerResponse

Arguments

:: Int
ccrsResponseStatus

-> CreateComputerResponse

Creates a value of CreateComputerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccrsComputer - A Computer object that represents the computer account.

ccrsResponseStatus - -- | The response status code.

data CreateComputerResponse

Contains the results for the CreateComputer operation.

See: createComputerResponse smart constructor.

Response Lenses

ccrsComputer :: Lens' CreateComputerResponse (Maybe Computer)

A Computer object that represents the computer account.

ccrsResponseStatus :: Lens' CreateComputerResponse Int

- | The response status code.