Network.AWS.DirectoryService.CreateConditionalForwarder

Description

Creates a conditional forwarder associated with your AWS directory. Conditional forwarders are required in order to set up a trust relationship with another domain. The conditional forwarder points to the trusted domain.

Creating a Request

createConditionalForwarder

Arguments

:: Text
ccfDirectoryId

-> Text
ccfRemoteDomainName

-> CreateConditionalForwarder

Creates a value of CreateConditionalForwarder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccfDirectoryId - The directory ID of the AWS directory for which you are creating the conditional forwarder.

ccfRemoteDomainName - The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.

ccfDNSIPAddrs - The IP addresses of the remote DNS server associated with RemoteDomainName.

data CreateConditionalForwarder

Initiates the creation of a conditional forwarder for your AWS Directory Service for Microsoft Active Directory. Conditional forwarders are required in order to set up a trust relationship with another domain.

See: createConditionalForwarder smart constructor.

Request Lenses

ccfDirectoryId :: Lens' CreateConditionalForwarder Text

The directory ID of the AWS directory for which you are creating the conditional forwarder.

ccfRemoteDomainName :: Lens' CreateConditionalForwarder Text

The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.

ccfDNSIPAddrs :: Lens' CreateConditionalForwarder [Text]

The IP addresses of the remote DNS server associated with RemoteDomainName.

Destructuring the Response

createConditionalForwarderResponse

Arguments

:: Int
ccfrsResponseStatus

-> CreateConditionalForwarderResponse

Creates a value of CreateConditionalForwarderResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccfrsResponseStatus - -- | The response status code.

data CreateConditionalForwarderResponse

The result of a CreateConditinalForwarder request.

See: createConditionalForwarderResponse smart constructor.

Response Lenses

ccfrsResponseStatus :: Lens' CreateConditionalForwarderResponse Int

- | The response status code.