Network.AWS.DirectoryService.CreateDirectory

Description

Creates a Simple AD directory.

Before you call CreateDirectory , ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference .

Creating a Request

createDirectory

Arguments

:: Text
cName

-> Text
cPassword

-> DirectorySize
cSize

-> CreateDirectory

Creates a value of CreateDirectory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cShortName - The short name of the directory, such as CORP .

cVPCSettings - A DirectoryVpcSettings object that contains additional information for the operation.

cDescription - A textual description for the directory.

cName - The fully qualified name for the directory, such as corp.example.com .

cPassword - The password for the directory administrator. The directory creation process creates a directory administrator account with the username Administrator and this password.

cSize - The size of the directory.

data CreateDirectory

Contains the inputs for the CreateDirectory operation.

See: createDirectory smart constructor.

Request Lenses

cShortName :: Lens' CreateDirectory (Maybe Text)

The short name of the directory, such as CORP .

cVPCSettings :: Lens' CreateDirectory (Maybe DirectoryVPCSettings)

A DirectoryVpcSettings object that contains additional information for the operation.

cDescription :: Lens' CreateDirectory (Maybe Text)

A textual description for the directory.

cName :: Lens' CreateDirectory Text

The fully qualified name for the directory, such as corp.example.com .

cPassword :: Lens' CreateDirectory Text

The password for the directory administrator. The directory creation process creates a directory administrator account with the username Administrator and this password.

cSize :: Lens' CreateDirectory DirectorySize

The size of the directory.

Destructuring the Response

createDirectoryResponse

Arguments

:: Int
crsResponseStatus

-> CreateDirectoryResponse

Creates a value of CreateDirectoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsDirectoryId - The identifier of the directory that was created.

crsResponseStatus - -- | The response status code.

data CreateDirectoryResponse

Contains the results of the CreateDirectory operation.

See: createDirectoryResponse smart constructor.

Response Lenses

crsDirectoryId :: Lens' CreateDirectoryResponse (Maybe Text)

The identifier of the directory that was created.

crsResponseStatus :: Lens' CreateDirectoryResponse Int

- | The response status code.