Network.AWS.DirectoryService.CreateMicrosoftAD

Description

Creates a Microsoft AD in the AWS cloud.

Before you call CreateMicrosoftAD , ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateMicrosoftAD operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference .

Creating a Request

createMicrosoftAD

Arguments

:: Text
cmadName

-> Text
cmadPassword

-> DirectoryVPCSettings
cmadVPCSettings

-> CreateMicrosoftAD

Creates a value of CreateMicrosoftAD with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmadShortName - The NetBIOS name for your domain. A short identifier for your domain, such as CORP . If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com .

cmadDescription - A textual description for the directory. This label will appear on the AWS console Directory Details page after the directory is created.

cmadName - The fully qualified domain name for the directory, such as corp.example.com . This name will resolve inside your VPC only. It does not need to be publicly resolvable.

cmadPassword - The password for the default administrative user named Admin .

cmadVPCSettings - Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

data CreateMicrosoftAD

Creates a Microsoft AD in the AWS cloud.

See: createMicrosoftAD smart constructor.

Request Lenses

cmadShortName :: Lens' CreateMicrosoftAD (Maybe Text)

The NetBIOS name for your domain. A short identifier for your domain, such as CORP . If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com .

cmadDescription :: Lens' CreateMicrosoftAD (Maybe Text)

A textual description for the directory. This label will appear on the AWS console Directory Details page after the directory is created.

cmadName :: Lens' CreateMicrosoftAD Text

The fully qualified domain name for the directory, such as corp.example.com . This name will resolve inside your VPC only. It does not need to be publicly resolvable.

cmadPassword :: Lens' CreateMicrosoftAD Text

The password for the default administrative user named Admin .

cmadVPCSettings :: Lens' CreateMicrosoftAD DirectoryVPCSettings

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

Destructuring the Response

createMicrosoftADResponse

Arguments

:: Int
cmadrsResponseStatus

-> CreateMicrosoftADResponse

Creates a value of CreateMicrosoftADResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmadrsDirectoryId - The identifier of the directory that was created.

cmadrsResponseStatus - -- | The response status code.

data CreateMicrosoftADResponse

Result of a CreateMicrosoftAD request.

See: createMicrosoftADResponse smart constructor.

Response Lenses

cmadrsDirectoryId :: Lens' CreateMicrosoftADResponse (Maybe Text)

The identifier of the directory that was created.

cmadrsResponseStatus :: Lens' CreateMicrosoftADResponse Int

- | The response status code.