Network.AWS.DirectoryService.CreateTrust

Description

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

Creating a Request

createTrust

Arguments

:: Text
ctDirectoryId

-> Text
ctRemoteDomainName

-> Text
ctTrustPassword

-> TrustDirection
ctTrustDirection

-> CreateTrust

Creates a value of CreateTrust with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctConditionalForwarderIPAddrs - The IP addresses of the remote DNS server associated with RemoteDomainName.

ctTrustType - The trust relationship type.

ctDirectoryId - The Directory ID of the Microsoft AD in the AWS cloud for which to establish the trust relationship.

ctRemoteDomainName - The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

ctTrustPassword - The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

ctTrustDirection - The direction of the trust relationship.

data CreateTrust

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

See: createTrust smart constructor.

Request Lenses

ctConditionalForwarderIPAddrs :: Lens' CreateTrust [Text]

The IP addresses of the remote DNS server associated with RemoteDomainName.

ctTrustType :: Lens' CreateTrust (Maybe TrustType)

The trust relationship type.

ctDirectoryId :: Lens' CreateTrust Text

The Directory ID of the Microsoft AD in the AWS cloud for which to establish the trust relationship.

ctRemoteDomainName :: Lens' CreateTrust Text

The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

ctTrustPassword :: Lens' CreateTrust Text

The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

ctTrustDirection :: Lens' CreateTrust TrustDirection

The direction of the trust relationship.

Destructuring the Response

createTrustResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateTrustResponse

Creates a value of CreateTrustResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsTrustId - A unique identifier for the trust relationship that was created.

ctrsResponseStatus - -- | The response status code.

data CreateTrustResponse

The result of a CreateTrust request.

See: createTrustResponse smart constructor.

Response Lenses

ctrsTrustId :: Lens' CreateTrustResponse (Maybe Text)

A unique identifier for the trust relationship that was created.

ctrsResponseStatus :: Lens' CreateTrustResponse Int

- | The response status code.