Network.AWS.DirectoryService.DeleteTrust

Description

Deletes an existing trust relationship between your Microsoft AD in the AWS cloud and an external domain.

Creating a Request

deleteTrust

Arguments

:: Text
dtTrustId

-> DeleteTrust

Creates a value of DeleteTrust with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtDeleteAssociatedConditionalForwarder - Delete a conditional forwarder as part of a DeleteTrustRequest.

dtTrustId - The Trust ID of the trust relationship to be deleted.

data DeleteTrust

Deletes the local side of an existing trust relationship between the Microsoft AD in the AWS cloud and the external domain.

See: deleteTrust smart constructor.

Request Lenses

dtDeleteAssociatedConditionalForwarder :: Lens' DeleteTrust (Maybe Bool)

Delete a conditional forwarder as part of a DeleteTrustRequest.

dtTrustId :: Lens' DeleteTrust Text

The Trust ID of the trust relationship to be deleted.

Destructuring the Response

deleteTrustResponse

Arguments

:: Int
dttrsResponseStatus

-> DeleteTrustResponse

Creates a value of DeleteTrustResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dttrsTrustId - The Trust ID of the trust relationship that was deleted.

dttrsResponseStatus - -- | The response status code.

data DeleteTrustResponse

The result of a DeleteTrust request.

See: deleteTrustResponse smart constructor.

Response Lenses

dttrsTrustId :: Lens' DeleteTrustResponse (Maybe Text)

The Trust ID of the trust relationship that was deleted.

dttrsResponseStatus :: Lens' DeleteTrustResponse Int

- | The response status code.