Network.AWS.DirectoryService.DescribeDirectories

Description

Obtains information about the directories that belong to this account.

You can retrieve information about specific directories by passing the directory identifiers in the DirectoryIds parameter. Otherwise, all directories that belong to the current account are returned.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeDirectoriesResult.NextToken member contains a token that you pass in the next call to DescribeDirectories to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

Creating a Request

describeDirectories :: DescribeDirectories

Creates a value of DescribeDirectories with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddNextToken - The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories . Pass null if this is the first call.

ddDirectoryIds - A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned. An empty list results in an InvalidParameterException being thrown.

ddLimit - The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

data DescribeDirectories

Contains the inputs for the DescribeDirectories operation.

See: describeDirectories smart constructor.

Request Lenses

ddNextToken :: Lens' DescribeDirectories (Maybe Text)

The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories . Pass null if this is the first call.

ddDirectoryIds :: Lens' DescribeDirectories [Text]

A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned. An empty list results in an InvalidParameterException being thrown.

ddLimit :: Lens' DescribeDirectories (Maybe Natural)

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

Destructuring the Response

describeDirectoriesResponse

Arguments

:: Int
ddrsResponseStatus

-> DescribeDirectoriesResponse

Creates a value of DescribeDirectoriesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsDirectoryDescriptions - The list of DirectoryDescription objects that were retrieved. It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

ddrsNextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDirectories to retrieve the next set of items.

ddrsResponseStatus - -- | The response status code.

data DescribeDirectoriesResponse

Contains the results of the DescribeDirectories operation.

See: describeDirectoriesResponse smart constructor.

Response Lenses

ddrsDirectoryDescriptions :: Lens' DescribeDirectoriesResponse [DirectoryDescription]

The list of DirectoryDescription objects that were retrieved. It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

ddrsNextToken :: Lens' DescribeDirectoriesResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDirectories to retrieve the next set of items.

ddrsResponseStatus :: Lens' DescribeDirectoriesResponse Int

- | The response status code.