Network.AWS.DirectoryService.DescribeDomainControllers

Description

Provides information about any domain controllers in your directory.

Creating a Request

describeDomainControllers

Arguments

:: Text
ddcDirectoryId

-> DescribeDomainControllers

Creates a value of DescribeDomainControllers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcNextToken - The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers . Pass null if this is the first call.

ddcDomainControllerIds - A list of identifiers for the domain controllers whose information will be provided.

ddcLimit - The maximum number of items to return.

ddcDirectoryId - Identifier of the directory for which to retrieve the domain controller information.

data DescribeDomainControllers

See: describeDomainControllers smart constructor.

Request Lenses

ddcNextToken :: Lens' DescribeDomainControllers (Maybe Text)

The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers . Pass null if this is the first call.

ddcDomainControllerIds :: Lens' DescribeDomainControllers [Text]

A list of identifiers for the domain controllers whose information will be provided.

ddcLimit :: Lens' DescribeDomainControllers (Maybe Natural)

The maximum number of items to return.

ddcDirectoryId :: Lens' DescribeDomainControllers Text

Identifier of the directory for which to retrieve the domain controller information.

Destructuring the Response

describeDomainControllersResponse

Arguments

:: Int
ddcrsResponseStatus

-> DescribeDomainControllersResponse

Creates a value of DescribeDomainControllersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddcrsNextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDomainControllers retrieve the next set of items.

ddcrsDomainControllers - List of the DomainController objects that were retrieved.

ddcrsResponseStatus - -- | The response status code.

data DescribeDomainControllersResponse

See: describeDomainControllersResponse smart constructor.

Response Lenses

ddcrsNextToken :: Lens' DescribeDomainControllersResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeDomainControllers retrieve the next set of items.

ddcrsDomainControllers :: Lens' DescribeDomainControllersResponse [DomainController]

List of the DomainController objects that were retrieved.

ddcrsResponseStatus :: Lens' DescribeDomainControllersResponse Int

- | The response status code.