Network.AWS.DirectoryService.DescribeTrusts

Description

Obtains information about the trust relationships for this account.

If no input parameters are provided, such as DirectoryId or TrustIds, this request describes all the trust relationships belonging to the account.

Creating a Request

describeTrusts :: DescribeTrusts

Creates a value of DescribeTrusts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtDirectoryId - The Directory ID of the AWS directory that is a part of the requested trust relationship.

dtNextToken - The DescribeTrustsResult.NextToken value from a previous call to DescribeTrusts . Pass null if this is the first call.

dtTrustIds - A list of identifiers of the trust relationships for which to obtain the information. If this member is null, all trust relationships that belong to the current account are returned. An empty list results in an InvalidParameterException being thrown.

dtLimit - The maximum number of objects to return.

data DescribeTrusts

Describes the trust relationships for a particular Microsoft AD in the AWS cloud. If no input parameters are are provided, such as directory ID or trust ID, this request describes all the trust relationships.

See: describeTrusts smart constructor.

Request Lenses

dtDirectoryId :: Lens' DescribeTrusts (Maybe Text)

The Directory ID of the AWS directory that is a part of the requested trust relationship.

dtNextToken :: Lens' DescribeTrusts (Maybe Text)

The DescribeTrustsResult.NextToken value from a previous call to DescribeTrusts . Pass null if this is the first call.

dtTrustIds :: Lens' DescribeTrusts [Text]

A list of identifiers of the trust relationships for which to obtain the information. If this member is null, all trust relationships that belong to the current account are returned. An empty list results in an InvalidParameterException being thrown.

dtLimit :: Lens' DescribeTrusts (Maybe Natural)

The maximum number of objects to return.

Destructuring the Response

describeTrustsResponse

Arguments

:: Int
dtrsResponseStatus

-> DescribeTrustsResponse

Creates a value of DescribeTrustsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsNextToken - If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeTrusts to retrieve the next set of items.

dtrsTrusts - The list of Trust objects that were retrieved. It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

dtrsResponseStatus - -- | The response status code.

data DescribeTrustsResponse

The result of a DescribeTrust request.

See: describeTrustsResponse smart constructor.

Response Lenses

dtrsNextToken :: Lens' DescribeTrustsResponse (Maybe Text)

If not null, more results are available. Pass this value for the NextToken parameter in a subsequent call to DescribeTrusts to retrieve the next set of items.

dtrsTrusts :: Lens' DescribeTrustsResponse [Trust]

The list of Trust objects that were retrieved. It is possible that this list contains less than the number of items specified in the Limit member of the request. This occurs if there are less than the requested number of items left to retrieve, or if the limitations of the operation have been exceeded.

dtrsResponseStatus :: Lens' DescribeTrustsResponse Int

- | The response status code.