Network.AWS.DirectoryService.DisableSSO

Description

Disables single-sign on for a directory.

Creating a Request

disableSSO

Arguments

:: Text
dssoDirectoryId

-> DisableSSO

Creates a value of DisableSSO with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssoUserName - The username of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. This account must have privileges to remove a service principal name. If the AD Connector service account does not have privileges to remove a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to disable single sign-on and are not stored by the service. The AD Connector service account is not changed.

dssoPassword - The password of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

dssoDirectoryId - The identifier of the directory for which to disable single-sign on.

data DisableSSO

Contains the inputs for the DisableSso operation.

See: disableSSO smart constructor.

Request Lenses

dssoUserName :: Lens' DisableSSO (Maybe Text)

The username of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. This account must have privileges to remove a service principal name. If the AD Connector service account does not have privileges to remove a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to disable single sign-on and are not stored by the service. The AD Connector service account is not changed.

dssoPassword :: Lens' DisableSSO (Maybe Text)

The password of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

dssoDirectoryId :: Lens' DisableSSO Text

The identifier of the directory for which to disable single-sign on.

Destructuring the Response

disableSSOResponse

Arguments

:: Int
dssorsResponseStatus

-> DisableSSOResponse

Creates a value of DisableSSOResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssorsResponseStatus - -- | The response status code.

data DisableSSOResponse

Contains the results of the DisableSso operation.

See: disableSSOResponse smart constructor.

Response Lenses

dssorsResponseStatus :: Lens' DisableSSOResponse Int

- | The response status code.