Network.AWS.DirectoryService.EnableSSO

Description

Enables single sign-on for a directory.

Creating a Request

enableSSO

Arguments

:: Text
esDirectoryId

-> EnableSSO

Creates a value of EnableSSO with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esUserName - The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name. If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

esPassword - The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

esDirectoryId - The identifier of the directory for which to enable single-sign on.

data EnableSSO

Contains the inputs for the EnableSso operation.

See: enableSSO smart constructor.

Request Lenses

esUserName :: Lens' EnableSSO (Maybe Text)

The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name. If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

esPassword :: Lens' EnableSSO (Maybe Text)

The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

esDirectoryId :: Lens' EnableSSO Text

The identifier of the directory for which to enable single-sign on.

Destructuring the Response

enableSSOResponse

Arguments

:: Int
esrsResponseStatus

-> EnableSSOResponse

Creates a value of EnableSSOResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esrsResponseStatus - -- | The response status code.

data EnableSSOResponse

Contains the results of the EnableSso operation.

See: enableSSOResponse smart constructor.

Response Lenses

esrsResponseStatus :: Lens' EnableSSOResponse Int

- | The response status code.