Network.AWS.DirectoryService.GetSnapshotLimits

Description

Obtains the manual snapshot limits for a directory.

Creating a Request

getSnapshotLimits

Arguments

:: Text
gslDirectoryId

-> GetSnapshotLimits

Creates a value of GetSnapshotLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gslDirectoryId - Contains the identifier of the directory to obtain the limits for.

data GetSnapshotLimits

Contains the inputs for the GetSnapshotLimits operation.

See: getSnapshotLimits smart constructor.

Request Lenses

gslDirectoryId :: Lens' GetSnapshotLimits Text

Contains the identifier of the directory to obtain the limits for.

Destructuring the Response

getSnapshotLimitsResponse

Arguments

:: Int
gslrsResponseStatus

-> GetSnapshotLimitsResponse

Creates a value of GetSnapshotLimitsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gslrsSnapshotLimits - A SnapshotLimits object that contains the manual snapshot limits for the specified directory.

gslrsResponseStatus - -- | The response status code.

data GetSnapshotLimitsResponse

Contains the results of the GetSnapshotLimits operation.

See: getSnapshotLimitsResponse smart constructor.

Response Lenses

gslrsSnapshotLimits :: Lens' GetSnapshotLimitsResponse (Maybe SnapshotLimits)

A SnapshotLimits object that contains the manual snapshot limits for the specified directory.

gslrsResponseStatus :: Lens' GetSnapshotLimitsResponse Int

- | The response status code.