Network.AWS.DirectoryService.Types

Description

Service Configuration

directoryService :: Service

API version 2015-04-16 of the Amazon Directory Service SDK configuration.

Errors

_DirectoryUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The specified directory is unavailable or could not be found.

_AuthenticationFailedException :: AsError a => Getting (First ServiceError) a ServiceError

An authentication error occurred.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters are not valid.

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The operation is not supported.

_EntityAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified entity already exists.

_DirectoryLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of directories in the region has been reached. You can use the GetDirectoryLimits operation to determine your directory limits in the region.

_IPRouteLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum allowed number of IP addresses was exceeded. The default limit is 100 IP address blocks.

_EntityDoesNotExistException :: AsError a => Getting (First ServiceError) a ServiceError

The specified entity could not be found.

_InsufficientPermissionsException :: AsError a => Getting (First ServiceError) a ServiceError

The account does not have sufficient permission to perform the operation.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken value is not valid.

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

An exception has occurred in AWS Directory Service.

_SnapshotLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of manual snapshots for the directory has been reached. You can use the GetSnapshotLimits operation to determine the snapshot limits for a directory.

_DomainControllerLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum allowed number of domain controllers per directory was exceeded. The default limit per directory is 20 domain controllers.

_TagLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum allowed number of tags was exceeded.

_ClientException :: AsError a => Getting (First ServiceError) a ServiceError

A client exception has occurred.

DirectorySize

data DirectorySize

Constructors

Large

Small

DirectoryStage

data DirectoryStage

Constructors

DSActive

DSCreated

DSCreating

DSDeleted

DSDeleting

DSFailed

DSImpaired

DSInoperable

DSRequested

DSRestoreFailed

DSRestoring

DirectoryType

data DirectoryType

Constructors

ADConnector

MicrosoftAD

SimpleAD

DomainControllerStatus

data DomainControllerStatus

Constructors

Active

Creating

Deleted

Deleting

Failed

Impaired

Restoring

IPRouteStatusMsg

data IPRouteStatusMsg

Constructors

AddFailed

Added

Adding

RemoveFailed

Removed

Removing

RadiusAuthenticationProtocol

data RadiusAuthenticationProtocol

Constructors

Chap

MsCHAPV1

MsCHAPV2

Pap

RadiusStatus

data RadiusStatus

Constructors

RSCompleted

RSCreating

RSFailed

ReplicationScope

data ReplicationScope

Constructors

Domain

SchemaExtensionStatus

data SchemaExtensionStatus

Constructors

SESCancelInProgress

SESCancelled

SESCompleted

SESCreatingSnapshot

SESFailed

SESInitializing

SESReplicating

SESRollbackInProgress

SESUpdatingSchema

SnapshotStatus

data SnapshotStatus

Constructors

SSCompleted

SSCreating

SSFailed

SnapshotType

data SnapshotType

Constructors

Auto

Manual

TopicStatus

data TopicStatus

Constructors

TDeleted

TFailed

TRegistered

TTopicNotFound

TrustDirection

data TrustDirection

Constructors

OneWayIncoming

OneWayOutgoing

TwoWay

TrustState

data TrustState

Constructors

TSCreated

TSCreating

TSDeleted

TSDeleting

TSFailed

TSVerified

TSVerifyFailed

TSVerifying

TrustType

data TrustType

Constructors

Forest

Attribute

data Attribute

Represents a named directory attribute.

See: attribute smart constructor.

attribute :: Attribute

Creates a value of Attribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aValue - The value of the attribute.

aName - The name of the attribute.

aValue :: Lens' Attribute (Maybe Text)

The value of the attribute.

aName :: Lens' Attribute (Maybe Text)

The name of the attribute.

Computer

data Computer

Contains information about a computer account in a directory.

See: computer smart constructor.

computer :: Computer

Creates a value of Computer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cComputerId - The identifier of the computer.

cComputerAttributes - An array of Attribute objects containing the LDAP attributes that belong to the computer account.

cComputerName - The computer name.

cComputerId :: Lens' Computer (Maybe Text)

The identifier of the computer.

cComputerAttributes :: Lens' Computer [Attribute]

An array of Attribute objects containing the LDAP attributes that belong to the computer account.

cComputerName :: Lens' Computer (Maybe Text)

The computer name.

ConditionalForwarder

data ConditionalForwarder

Points to a remote domain with which you are setting up a trust relationship. Conditional forwarders are required in order to set up a trust relationship with another domain.

See: conditionalForwarder smart constructor.

conditionalForwarder :: ConditionalForwarder

Creates a value of ConditionalForwarder with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfDNSIPAddrs - The IP addresses of the remote DNS server associated with RemoteDomainName. This is the IP address of the DNS server that your conditional forwarder points to.

cfRemoteDomainName - The fully qualified domain name (FQDN) of the remote domains pointed to by the conditional forwarder.

cfReplicationScope - The replication scope of the conditional forwarder. The only allowed value is Domain , which will replicate the conditional forwarder to all of the domain controllers for your AWS directory.

cfDNSIPAddrs :: Lens' ConditionalForwarder [Text]

The IP addresses of the remote DNS server associated with RemoteDomainName. This is the IP address of the DNS server that your conditional forwarder points to.

cfRemoteDomainName :: Lens' ConditionalForwarder (Maybe Text)

The fully qualified domain name (FQDN) of the remote domains pointed to by the conditional forwarder.

cfReplicationScope :: Lens' ConditionalForwarder (Maybe ReplicationScope)

The replication scope of the conditional forwarder. The only allowed value is Domain , which will replicate the conditional forwarder to all of the domain controllers for your AWS directory.

DirectoryConnectSettings

data DirectoryConnectSettings

Contains information for the ConnectDirectory operation when an AD Connector directory is being created.

See: directoryConnectSettings smart constructor.

directoryConnectSettings

Arguments

:: Text
dcsVPCId

-> Text
dcsCustomerUserName

-> DirectoryConnectSettings

Creates a value of DirectoryConnectSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsVPCId - The identifier of the VPC in which the AD Connector is created.

dcsSubnetIds - A list of subnet identifiers in the VPC in which the AD Connector is created.

dcsCustomerDNSIPs - A list of one or more IP addresses of DNS servers or domain controllers in the on-premises directory.

dcsCustomerUserName - The username of an account in the on-premises directory that is used to connect to the directory. This account must have the following privileges: * Read users and groups * Create computer objects * Join computers to the domain

dcsVPCId :: Lens' DirectoryConnectSettings Text

The identifier of the VPC in which the AD Connector is created.

dcsSubnetIds :: Lens' DirectoryConnectSettings [Text]

A list of subnet identifiers in the VPC in which the AD Connector is created.

dcsCustomerDNSIPs :: Lens' DirectoryConnectSettings [Text]

A list of one or more IP addresses of DNS servers or domain controllers in the on-premises directory.

dcsCustomerUserName :: Lens' DirectoryConnectSettings Text

The username of an account in the on-premises directory that is used to connect to the directory. This account must have the following privileges: * Read users and groups * Create computer objects * Join computers to the domain

DirectoryConnectSettingsDescription

data DirectoryConnectSettingsDescription

Contains information about an AD Connector directory.

See: directoryConnectSettingsDescription smart constructor.

directoryConnectSettingsDescription :: DirectoryConnectSettingsDescription

Creates a value of DirectoryConnectSettingsDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcsdCustomerUserName - The username of the service account in the on-premises directory.

dcsdSubnetIds - A list of subnet identifiers in the VPC that the AD connector is in.

dcsdVPCId - The identifier of the VPC that the AD Connector is in.

dcsdSecurityGroupId - The security group identifier for the AD Connector directory.

dcsdConnectIPs - The IP addresses of the AD Connector servers.

dcsdAvailabilityZones - A list of the Availability Zones that the directory is in.

dcsdCustomerUserName :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The username of the service account in the on-premises directory.

dcsdSubnetIds :: Lens' DirectoryConnectSettingsDescription [Text]

A list of subnet identifiers in the VPC that the AD connector is in.

dcsdVPCId :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The identifier of the VPC that the AD Connector is in.

dcsdSecurityGroupId :: Lens' DirectoryConnectSettingsDescription (Maybe Text)

The security group identifier for the AD Connector directory.

dcsdConnectIPs :: Lens' DirectoryConnectSettingsDescription [Text]

The IP addresses of the AD Connector servers.

dcsdAvailabilityZones :: Lens' DirectoryConnectSettingsDescription [Text]

A list of the Availability Zones that the directory is in.

DirectoryDescription

data DirectoryDescription

Contains information about an AWS Directory Service directory.

See: directoryDescription smart constructor.

directoryDescription :: DirectoryDescription

Creates a value of DirectoryDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddRadiusStatus - The status of the RADIUS MFA server connection.

ddStage - The current stage of the directory.

ddDirectoryId - The directory identifier.

ddAccessURL - The access URL for the directory, such as http://<alias>.awsapps.com . If no alias has been created for the directory, alias is the directory identifier, such as d-XXXXXXXXXX .

ddShortName - The short name of the directory.

ddSize - The directory size.

ddDesiredNumberOfDomainControllers - The desired number of domain controllers in the directory if the directory is Microsoft AD.

ddRadiusSettings - A RadiusSettings object that contains information about the RADIUS server configured for this directory.

ddLaunchTime - Specifies when the directory was created.

ddAlias - The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX .

ddName - The fully-qualified name of the directory.

ddStageLastUpdatedDateTime - The date and time that the stage was last updated.

ddSSOEnabled - Indicates if single-sign on is enabled for the directory. For more information, see EnableSso and DisableSso .

ddDNSIPAddrs - The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in the on-premises directory to which the AD Connector is connected.

ddVPCSettings - A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed AD directory.

ddType - The directory size.

ddStageReason - Additional information about the directory stage.

ddConnectSettings - A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

ddDescription - The textual description for the directory.

ddRadiusStatus :: Lens' DirectoryDescription (Maybe RadiusStatus)

The status of the RADIUS MFA server connection.

ddStage :: Lens' DirectoryDescription (Maybe DirectoryStage)

The current stage of the directory.

ddDirectoryId :: Lens' DirectoryDescription (Maybe Text)

The directory identifier.

ddAccessURL :: Lens' DirectoryDescription (Maybe Text)

The access URL for the directory, such as http://<alias>.awsapps.com . If no alias has been created for the directory, alias is the directory identifier, such as d-XXXXXXXXXX .

ddShortName :: Lens' DirectoryDescription (Maybe Text)

The short name of the directory.

ddSize :: Lens' DirectoryDescription (Maybe DirectorySize)

The directory size.

ddDesiredNumberOfDomainControllers :: Lens' DirectoryDescription (Maybe Natural)

The desired number of domain controllers in the directory if the directory is Microsoft AD.

ddRadiusSettings :: Lens' DirectoryDescription (Maybe RadiusSettings)

A RadiusSettings object that contains information about the RADIUS server configured for this directory.

ddLaunchTime :: Lens' DirectoryDescription (Maybe UTCTime)

Specifies when the directory was created.

ddAlias :: Lens' DirectoryDescription (Maybe Text)

The alias for the directory. If no alias has been created for the directory, the alias is the directory identifier, such as d-XXXXXXXXXX .

ddName :: Lens' DirectoryDescription (Maybe Text)

The fully-qualified name of the directory.

ddStageLastUpdatedDateTime :: Lens' DirectoryDescription (Maybe UTCTime)

The date and time that the stage was last updated.

ddSSOEnabled :: Lens' DirectoryDescription (Maybe Bool)

Indicates if single-sign on is enabled for the directory. For more information, see EnableSso and DisableSso .

ddDNSIPAddrs :: Lens' DirectoryDescription [Text]

The IP addresses of the DNS servers for the directory. For a Simple AD or Microsoft AD directory, these are the IP addresses of the Simple AD or Microsoft AD directory servers. For an AD Connector directory, these are the IP addresses of the DNS servers or domain controllers in the on-premises directory to which the AD Connector is connected.

ddVPCSettings :: Lens' DirectoryDescription (Maybe DirectoryVPCSettingsDescription)

A DirectoryVpcSettingsDescription object that contains additional information about a directory. This member is only present if the directory is a Simple AD or Managed AD directory.

ddType :: Lens' DirectoryDescription (Maybe DirectoryType)

The directory size.

ddStageReason :: Lens' DirectoryDescription (Maybe Text)

Additional information about the directory stage.

ddConnectSettings :: Lens' DirectoryDescription (Maybe DirectoryConnectSettingsDescription)

A DirectoryConnectSettingsDescription object that contains additional information about an AD Connector directory. This member is only present if the directory is an AD Connector directory.

ddDescription :: Lens' DirectoryDescription (Maybe Text)

The textual description for the directory.

DirectoryLimits

data DirectoryLimits

Contains directory limit information for a region.

See: directoryLimits smart constructor.

directoryLimits :: DirectoryLimits

Creates a value of DirectoryLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlConnectedDirectoriesCurrentCount - The current number of connected directories in the region.

dlCloudOnlyMicrosoftADLimitReached - Indicates if the Microsoft AD directory limit has been reached.

dlConnectedDirectoriesLimit - The maximum number of connected directories allowed in the region.

dlConnectedDirectoriesLimitReached - Indicates if the connected directory limit has been reached.

dlCloudOnlyMicrosoftADLimit - The maximum number of Microsoft AD directories allowed in the region.

dlCloudOnlyDirectoriesLimit - The maximum number of cloud directories allowed in the region.

dlCloudOnlyDirectoriesCurrentCount - The current number of cloud directories in the region.

dlCloudOnlyDirectoriesLimitReached - Indicates if the cloud directory limit has been reached.

dlCloudOnlyMicrosoftADCurrentCount - The current number of Microsoft AD directories in the region.

dlConnectedDirectoriesCurrentCount :: Lens' DirectoryLimits (Maybe Natural)

The current number of connected directories in the region.

dlCloudOnlyMicrosoftADLimitReached :: Lens' DirectoryLimits (Maybe Bool)

Indicates if the Microsoft AD directory limit has been reached.

dlConnectedDirectoriesLimit :: Lens' DirectoryLimits (Maybe Natural)

The maximum number of connected directories allowed in the region.

dlConnectedDirectoriesLimitReached :: Lens' DirectoryLimits (Maybe Bool)

Indicates if the connected directory limit has been reached.

dlCloudOnlyMicrosoftADLimit :: Lens' DirectoryLimits (Maybe Natural)

The maximum number of Microsoft AD directories allowed in the region.

dlCloudOnlyDirectoriesLimit :: Lens' DirectoryLimits (Maybe Natural)

The maximum number of cloud directories allowed in the region.

dlCloudOnlyDirectoriesCurrentCount :: Lens' DirectoryLimits (Maybe Natural)

The current number of cloud directories in the region.

dlCloudOnlyDirectoriesLimitReached :: Lens' DirectoryLimits (Maybe Bool)

Indicates if the cloud directory limit has been reached.

dlCloudOnlyMicrosoftADCurrentCount :: Lens' DirectoryLimits (Maybe Natural)

The current number of Microsoft AD directories in the region.

DirectoryVPCSettings

data DirectoryVPCSettings

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

See: directoryVPCSettings smart constructor.

directoryVPCSettings

Arguments

:: Text
dvsVPCId

-> DirectoryVPCSettings

Creates a value of DirectoryVPCSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvsVPCId - The identifier of the VPC in which to create the directory.

dvsSubnetIds - The identifiers of the subnets for the directory servers. The two subnets must be in different Availability Zones. AWS Directory Service creates a directory server and a DNS server in each of these subnets.

dvsVPCId :: Lens' DirectoryVPCSettings Text

The identifier of the VPC in which to create the directory.

dvsSubnetIds :: Lens' DirectoryVPCSettings [Text]

The identifiers of the subnets for the directory servers. The two subnets must be in different Availability Zones. AWS Directory Service creates a directory server and a DNS server in each of these subnets.

DirectoryVPCSettingsDescription

data DirectoryVPCSettingsDescription

Contains information about the directory.

See: directoryVPCSettingsDescription smart constructor.

directoryVPCSettingsDescription :: DirectoryVPCSettingsDescription

Creates a value of DirectoryVPCSettingsDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvsdSubnetIds - The identifiers of the subnets for the directory servers.

dvsdVPCId - The identifier of the VPC that the directory is in.

dvsdSecurityGroupId - The security group identifier for the directory. If the directory was created before 812014, this is the identifier of the directory members security group that was created when the directory was created. If the directory was created after this date, this value is null.

dvsdAvailabilityZones - The list of Availability Zones that the directory is in.

dvsdSubnetIds :: Lens' DirectoryVPCSettingsDescription [Text]

The identifiers of the subnets for the directory servers.

dvsdVPCId :: Lens' DirectoryVPCSettingsDescription (Maybe Text)

The identifier of the VPC that the directory is in.

dvsdSecurityGroupId :: Lens' DirectoryVPCSettingsDescription (Maybe Text)

The security group identifier for the directory. If the directory was created before 812014, this is the identifier of the directory members security group that was created when the directory was created. If the directory was created after this date, this value is null.

dvsdAvailabilityZones :: Lens' DirectoryVPCSettingsDescription [Text]

The list of Availability Zones that the directory is in.

DomainController

data DomainController

Contains information about the domain controllers for a specified directory.

See: domainController smart constructor.

domainController :: DomainController

Creates a value of DomainController with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcStatus - The status of the domain controller.

dcDirectoryId - Identifier of the directory where the domain controller resides.

dcVPCId - The identifier of the VPC that contains the domain controller.

dcLaunchTime - Specifies when the domain controller was created.

dcSubnetId - Identifier of the subnet in the VPC that contains the domain controller.

dcAvailabilityZone - The Availability Zone where the domain controller is located.

dcStatusLastUpdatedDateTime - The date and time that the status was last updated.

dcStatusReason - A description of the domain controller state.

dcDNSIPAddr - The IP address of the domain controller.

dcDomainControllerId - Identifies a specific domain controller in the directory.

dcStatus :: Lens' DomainController (Maybe DomainControllerStatus)

The status of the domain controller.

dcDirectoryId :: Lens' DomainController (Maybe Text)

Identifier of the directory where the domain controller resides.

dcVPCId :: Lens' DomainController (Maybe Text)

The identifier of the VPC that contains the domain controller.

dcLaunchTime :: Lens' DomainController (Maybe UTCTime)

Specifies when the domain controller was created.

dcSubnetId :: Lens' DomainController (Maybe Text)

Identifier of the subnet in the VPC that contains the domain controller.

dcAvailabilityZone :: Lens' DomainController (Maybe Text)

The Availability Zone where the domain controller is located.

dcStatusLastUpdatedDateTime :: Lens' DomainController (Maybe UTCTime)

The date and time that the status was last updated.

dcStatusReason :: Lens' DomainController (Maybe Text)

A description of the domain controller state.

dcDNSIPAddr :: Lens' DomainController (Maybe Text)

The IP address of the domain controller.

dcDomainControllerId :: Lens' DomainController (Maybe Text)

Identifies a specific domain controller in the directory.

EventTopic

data EventTopic

Information about SNS topic and AWS Directory Service directory associations.

See: eventTopic smart constructor.

eventTopic :: EventTopic

Creates a value of EventTopic with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etStatus - The topic registration status.

etDirectoryId - The Directory ID of an AWS Directory Service directory that will publish status messages to an SNS topic.

etTopicName - The name of an AWS SNS topic the receives status messages from the directory.

etTopicARN - The SNS topic ARN (Amazon Resource Name).

etCreatedDateTime - The date and time of when you associated your directory with the SNS topic.

etStatus :: Lens' EventTopic (Maybe TopicStatus)

The topic registration status.

etDirectoryId :: Lens' EventTopic (Maybe Text)

The Directory ID of an AWS Directory Service directory that will publish status messages to an SNS topic.

etTopicName :: Lens' EventTopic (Maybe Text)

The name of an AWS SNS topic the receives status messages from the directory.

etTopicARN :: Lens' EventTopic (Maybe Text)

The SNS topic ARN (Amazon Resource Name).

etCreatedDateTime :: Lens' EventTopic (Maybe UTCTime)

The date and time of when you associated your directory with the SNS topic.

IPRoute

data IPRoute

IP address block. This is often the address block of the DNS server used for your on-premises domain.

See: ipRoute smart constructor.

ipRoute :: IPRoute

Creates a value of IPRoute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irCidrIP - IP address block using CIDR format, for example 10.0.0.024. This is often the address block of the DNS server used for your on-premises domain. For a single IP address use a CIDR address block with 32. For example 10.0.0.0/32.

irDescription - Description of the address block.

irCidrIP :: Lens' IPRoute (Maybe Text)

IP address block using CIDR format, for example 10.0.0.024. This is often the address block of the DNS server used for your on-premises domain. For a single IP address use a CIDR address block with 32. For example 10.0.0.0/32.

irDescription :: Lens' IPRoute (Maybe Text)

Description of the address block.

IPRouteInfo

data IPRouteInfo

Information about one or more IP address blocks.

See: ipRouteInfo smart constructor.

ipRouteInfo :: IPRouteInfo

Creates a value of IPRouteInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iriDirectoryId - Identifier (ID) of the directory associated with the IP addresses.

iriIPRouteStatusReason - The reason for the IpRouteStatusMsg.

iriAddedDateTime - The date and time the address block was added to the directory.

iriCidrIP - IP address block in the IpRoute .

iriIPRouteStatusMsg - The status of the IP address block.

iriDescription - Description of the IpRouteInfo .

iriDirectoryId :: Lens' IPRouteInfo (Maybe Text)

Identifier (ID) of the directory associated with the IP addresses.

iriIPRouteStatusReason :: Lens' IPRouteInfo (Maybe Text)

The reason for the IpRouteStatusMsg.

iriAddedDateTime :: Lens' IPRouteInfo (Maybe UTCTime)

The date and time the address block was added to the directory.

iriCidrIP :: Lens' IPRouteInfo (Maybe Text)

IP address block in the IpRoute .

iriIPRouteStatusMsg :: Lens' IPRouteInfo (Maybe IPRouteStatusMsg)

The status of the IP address block.

iriDescription :: Lens' IPRouteInfo (Maybe Text)

Description of the IpRouteInfo .

RadiusSettings

data RadiusSettings

Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

See: radiusSettings smart constructor.

radiusSettings :: RadiusSettings

Creates a value of RadiusSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsDisplayLabel - Not currently used.

rsRadiusRetries - The maximum number of times that communication with the RADIUS server is attempted.

rsAuthenticationProtocol - The protocol specified for your RADIUS endpoints.

rsRadiusServers - An array of strings that contains the IP addresses of the RADIUS server endpoints, or the IP addresses of your RADIUS server load balancer.

rsUseSameUsername - Not currently used.

rsSharedSecret - Not currently used.

rsRadiusTimeout - The amount of time, in seconds, to wait for the RADIUS server to respond.

rsRadiusPort - The port that your RADIUS server is using for communications. Your on-premises network must allow inbound traffic over this port from the AWS Directory Service servers.

rsDisplayLabel :: Lens' RadiusSettings (Maybe Text)

Not currently used.

rsRadiusRetries :: Lens' RadiusSettings (Maybe Natural)

The maximum number of times that communication with the RADIUS server is attempted.

rsAuthenticationProtocol :: Lens' RadiusSettings (Maybe RadiusAuthenticationProtocol)

The protocol specified for your RADIUS endpoints.

rsRadiusServers :: Lens' RadiusSettings [Text]

An array of strings that contains the IP addresses of the RADIUS server endpoints, or the IP addresses of your RADIUS server load balancer.

rsUseSameUsername :: Lens' RadiusSettings (Maybe Bool)

Not currently used.

rsSharedSecret :: Lens' RadiusSettings (Maybe Text)

Not currently used.

rsRadiusTimeout :: Lens' RadiusSettings (Maybe Natural)

The amount of time, in seconds, to wait for the RADIUS server to respond.

rsRadiusPort :: Lens' RadiusSettings (Maybe Natural)

The port that your RADIUS server is using for communications. Your on-premises network must allow inbound traffic over this port from the AWS Directory Service servers.

SchemaExtensionInfo

data SchemaExtensionInfo

Information about a schema extension.

See: schemaExtensionInfo smart constructor.

schemaExtensionInfo :: SchemaExtensionInfo

Creates a value of SchemaExtensionInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

seiDirectoryId - The identifier of the directory to which the schema extension is applied.

seiSchemaExtensionId - The identifier of the schema extension.

seiSchemaExtensionStatusReason - The reason for the SchemaExtensionStatus .

seiSchemaExtensionStatus - The current status of the schema extension.

seiDescription - A description of the schema extension.

seiEndDateTime - The date and time that the schema extension was completed.

seiStartDateTime - The date and time that the schema extension started being applied to the directory.

seiDirectoryId :: Lens' SchemaExtensionInfo (Maybe Text)

The identifier of the directory to which the schema extension is applied.

seiSchemaExtensionId :: Lens' SchemaExtensionInfo (Maybe Text)

The identifier of the schema extension.

seiSchemaExtensionStatusReason :: Lens' SchemaExtensionInfo (Maybe Text)

The reason for the SchemaExtensionStatus .

seiSchemaExtensionStatus :: Lens' SchemaExtensionInfo (Maybe SchemaExtensionStatus)

The current status of the schema extension.

seiDescription :: Lens' SchemaExtensionInfo (Maybe Text)

A description of the schema extension.

seiEndDateTime :: Lens' SchemaExtensionInfo (Maybe UTCTime)

The date and time that the schema extension was completed.

seiStartDateTime :: Lens' SchemaExtensionInfo (Maybe UTCTime)

The date and time that the schema extension started being applied to the directory.

Snapshot

data Snapshot

Describes a directory snapshot.

See: snapshot smart constructor.

snapshot :: Snapshot

Creates a value of Snapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sStatus - The snapshot status.

sDirectoryId - The directory identifier.

sStartTime - The date and time that the snapshot was taken.

sName - The descriptive name of the snapshot.

sType - The snapshot type.

sSnapshotId - The snapshot identifier.

sStatus :: Lens' Snapshot (Maybe SnapshotStatus)

The snapshot status.

sDirectoryId :: Lens' Snapshot (Maybe Text)

The directory identifier.

sStartTime :: Lens' Snapshot (Maybe UTCTime)

The date and time that the snapshot was taken.

sName :: Lens' Snapshot (Maybe Text)

The descriptive name of the snapshot.

sType :: Lens' Snapshot (Maybe SnapshotType)

The snapshot type.

sSnapshotId :: Lens' Snapshot (Maybe Text)

The snapshot identifier.

SnapshotLimits

data SnapshotLimits

Contains manual snapshot limit information for a directory.

See: snapshotLimits smart constructor.

snapshotLimits :: SnapshotLimits

Creates a value of SnapshotLimits with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slManualSnapshotsLimitReached - Indicates if the manual snapshot limit has been reached.

slManualSnapshotsCurrentCount - The current number of manual snapshots of the directory.

slManualSnapshotsLimit - The maximum number of manual snapshots allowed.

slManualSnapshotsLimitReached :: Lens' SnapshotLimits (Maybe Bool)

Indicates if the manual snapshot limit has been reached.

slManualSnapshotsCurrentCount :: Lens' SnapshotLimits (Maybe Natural)

The current number of manual snapshots of the directory.

slManualSnapshotsLimit :: Lens' SnapshotLimits (Maybe Natural)

The maximum number of manual snapshots allowed.

Tag

data Tag

Metadata assigned to a directory consisting of a key-value pair.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - Required name of the tag. The string value can be Unicode characters and cannot be prefixed with "aws:". The string can contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagValue - The optional value of the tag. The string value can be Unicode characters. The string can contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagKey :: Lens' Tag Text

Required name of the tag. The string value can be Unicode characters and cannot be prefixed with "aws:". The string can contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagValue :: Lens' Tag Text

The optional value of the tag. The string value can be Unicode characters. The string can contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

Trust

data Trust

Describes a trust relationship between an Microsoft AD in the AWS cloud and an external domain.

See: trust smart constructor.

trust :: Trust

Creates a value of Trust with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tDirectoryId - The Directory ID of the AWS directory involved in the trust relationship.

tTrustState - The trust relationship state.

tLastUpdatedDateTime - The date and time that the trust relationship was last updated.

tTrustDirection - The trust relationship direction.

tStateLastUpdatedDateTime - The date and time that the TrustState was last updated.

tTrustType - The trust relationship type.

tTrustStateReason - The reason for the TrustState.

tRemoteDomainName - The Fully Qualified Domain Name (FQDN) of the external domain involved in the trust relationship.

tTrustId - The unique ID of the trust relationship.

tCreatedDateTime - The date and time that the trust relationship was created.

tDirectoryId :: Lens' Trust (Maybe Text)

The Directory ID of the AWS directory involved in the trust relationship.

tTrustState :: Lens' Trust (Maybe TrustState)

The trust relationship state.

tLastUpdatedDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the trust relationship was last updated.

tTrustDirection :: Lens' Trust (Maybe TrustDirection)

The trust relationship direction.

tStateLastUpdatedDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the TrustState was last updated.

tTrustType :: Lens' Trust (Maybe TrustType)

The trust relationship type.

tTrustStateReason :: Lens' Trust (Maybe Text)

The reason for the TrustState.

tRemoteDomainName :: Lens' Trust (Maybe Text)

The Fully Qualified Domain Name (FQDN) of the external domain involved in the trust relationship.

tTrustId :: Lens' Trust (Maybe Text)

The unique ID of the trust relationship.

tCreatedDateTime :: Lens' Trust (Maybe UTCTime)

The date and time that the trust relationship was created.