Network.AWS.DirectoryService.UpdateNumberOfDomainControllers

Description

Adds or removes domain controllers to or from the directory. Based on the difference between current value and new value (provided through this API call), domain controllers will be added or removed. It may take up to 45 minutes for any new domain controllers to become fully active once the requested number of domain controllers is updated. During this time, you cannot make another update request.

Creating a Request

updateNumberOfDomainControllers

Arguments

:: Text
unodcDirectoryId

-> Natural
unodcDesiredNumber

-> UpdateNumberOfDomainControllers

Creates a value of UpdateNumberOfDomainControllers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unodcDirectoryId - Identifier of the directory to which the domain controllers will be added or removed.

unodcDesiredNumber - The number of domain controllers desired in the directory.

data UpdateNumberOfDomainControllers

See: updateNumberOfDomainControllers smart constructor.

Request Lenses

unodcDirectoryId :: Lens' UpdateNumberOfDomainControllers Text

Identifier of the directory to which the domain controllers will be added or removed.

unodcDesiredNumber :: Lens' UpdateNumberOfDomainControllers Natural

The number of domain controllers desired in the directory.

Destructuring the Response

updateNumberOfDomainControllersResponse

Arguments

:: Int
unodcrsResponseStatus

-> UpdateNumberOfDomainControllersResponse

Creates a value of UpdateNumberOfDomainControllersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unodcrsResponseStatus - -- | The response status code.

data UpdateNumberOfDomainControllersResponse

See: updateNumberOfDomainControllersResponse smart constructor.

Response Lenses

unodcrsResponseStatus :: Lens' UpdateNumberOfDomainControllersResponse Int

- | The response status code.